ASPCB: Rev 1.0f
Rerouted with DRDY select lines to I2C GPIO expander. This means that we
can check all of the DRDY lines with one interrupt, so long as we poll
the input expander after recieving each interrupt. This allows us to
switch to interrupt driven programming, which should enhance the data
collection speed while reducing CPU use. To accomodate the extra lines,
we have switched to gridded ground style routing.